backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
easyocr==1.7.1
torch==2.5.1
pillow==10.4.0
numpy==2.1.2
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
